Fix issue with styleCustomisation parameter
Refactor the feedback components to allow style declarations to be
passed as strings in the `style-customisation` attribute.

Make the StyleCustomisation interface properties optional in order to
prevent passing the `toast` configuration to an alert component and
vice-versa.

Set a default empty object for the style-customisation attribute in order
to prevent `undefined` errors